Code cleanup for player & better handling video loading error detection
