+++ materialious/src/lib/i18n/locales/en.json
@@ -92,7 +92,11 @@
 		"unableToLoadComments": "Unable to load comments",
 		"addToPlaylist": "Add to playlist",
 		"youtubeJsFallBack": "Local video fallback being used",
-		"defaultLanguage": "Default Language"
+		"defaultLanguage": "Default Language",
+		"youtubeJsLoading": "Video failed to load, attempting to reload with local video fallback.",
+		"retryText": "Video failed to load, how would you like to continue?",
+		"enableYoutubejsTemp": "Use local video fallback once",
+		"enableYoutubejsPerm": "Always allow local video fallback"
 	},
 	"layout": {
 		"interface": "Interface",
